feat: split gettext
Split the gettext translation into the text part and the editorial part that contains the links. This way the text part can be made available for translation on Weblate, whereas the editorial part can be kept in the repository for the technical details. This way translators are no longer able to make translation errors on the image links and can achieve a 100% translation score.

Existing translations need to be migrated. Necessary commands are in the comments.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>