feat(docbook/Makefile): quick-check command
Add quick-check command that generates all output docbook files and runs xmllint on them. This ensures the translations can be merged and the resulting docbook is valid XML. It does not validate the structure, or the subsequent processing of this docbook file into outputs like the Scribus document.

This test requires make, itstool and xmllint. This would be good to check translations.

Related to https://git.fsfe.org/FSFE/ada-zangemann/issues/302

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>