fix: prevent multiple EN targets in docbook Makefile
Resolve issue reported by Marco Ciampa of the Makefile warning about multiple targets for 'EN'.

The source of the problem was that for an 'EN' output, the necessary en.po files would be generated.
The en.po would be scanned as a translation, to which EN was added explicitly to have all available languages.
In that sense there was no issue, other than warnings.

This commit addresses the issue by ensuring that EN is not part of the translations, and therefore only once in the available languages.
Furthermore the generated en.po files added to the gitignore to prevent accidentally committing a file that is automatically generated.

An idea for the future would be to do all gettext handling in the build directory instead of in the source directory, to avoid these issues.
Then again, this setup makes it easier to create new language files when needed.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>