feat(docbook/scribus): align headings and text areas
Focus on text in later sections of the book. Equal distance between the text and headings on page with license text.

The text block below the ice creams is now aligned to the base page structure. It was aligned differently for the Italian and Spanish version, slightly to the right.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>