feat: screen pdf export
Add the option to export Scribus to a pdf file without crop marks or bleeds, so it is targeted for screen viewing.
In this way only a single template is needed for both print and screen targets.

As it seems impossible to parse arguments in the pdf file, it makes sense to have a separate file for this target.
This pattern is carried through in the .sh file for now.

Usage from docbook folder:

```
make build/Ada_Zangemann-print-headings-text.da.screen.pdf
```

Note that the output does not include a cover. Adding the cover might require a separate cover template that is combined, or by having a separate template altogether.

Addresses https://git.fsfe.org/FSFE/ada-zangemann/issues/353

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>