fix(docbook/xsl): Work around Pandoc title bug
Ensure that the main <info> element is at the end of the docbook document, so that Pandoc uses the correct title as the book title.

It does this by first skipping over the <info> element and than copying it explicitly at the end.

This is a workaround for issue https://git.fsfe.org/FSFE/ada-zangemann/issues/384 reported upstream as https://github.com/jgm/pandoc/issues/11300

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>