chore(docbook/templates): realign text frame on page 11
Better algin to German and English books. The text frame on page 11 is less wide and the right edge follows the skew of the drawing.
This change ensures a skewed line from the top of the page, and solves the situation where it only began after 3 lines, hiding the skew.

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>