nn: fix almost all fuzzies
I used poedit for this, and it has two modes: either unwrap everything
to very long lines, or wrap everything to 79 lines.  To keep the diff
readable, I chose the latter, although I prefer to work with long lines
in reality.  this did however re-wrap some translations which were not
actually touched.