feat(docbook): confidential overlay watermark
Transparent overlay with text 'Confidential Preview', intended for screen pdfs.

Intended for the use-case of sharing a pdf for review that is not intended for broader sharing.
Note that the CC-BY-SA lice does allow sharing, so it is a kind ask, not something to enforce.

Implemented using Ghostscript. Often pdftk is recommended to overlay a pdf or PostScript file onto an existing pdf, but adopting pdftk would introduce another project dependency.
By leveraging Ghostscript transparency extension in PostScript, combined with a 'hack' to write in PostScript onto existing pages, the overlay can be created.

Ghostscript returns an error when running, which is still to be addressed:

```
Error: /rangecheck in .installpagedevice
Operand stack:
   --nostringval--   58   0   --nostringval--   --dict:237/249(ro)(L)--   --nostringval--   --nostringval--   false
Execution stack:
   %interp_exit   .runexec2   --nostringval--   restore   --nostringval--   2   %stopped_push   --nostringval--   restore   restore   false   1   %stopped_push   1949   1   3   %oparray_pop   1948   1   3   %oparray_pop   1933   1   3   %oparray_pop   1934   1   3   %oparray_pop   restore   restore   1868   4   4   %oparray_pop   restore   1849   4   4   %oparray_pop   restore   restore
Dictionary stack:
   --dict:766/1123(ro)(G)--   --dict:0/20(G)--   --dict:88/200(L)--   --dict:2/10(L)--
Current allocation mode is local
GPL Ghostscript 10.05.1: Unrecoverable error, exit code 1
GPL Ghostscript 10.05.1: Page object was reserved for an Annotation destination, but no such page was drawn, annotation in output will be invalid.
```

The resulting pdf seems to be fine though.
For now the error state is ignored using `|| true`.

Font size and position is designed for 210x210mm format for screen pdf.
In future the script can be extended in case more variants are needed.

Related to issue: https://git.fsfe.org/FSFE/ada-zangemann/issues/425

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>