feat(docbook/xsl/scribus): strip-first-character processing-configuration
Add a mechanism to the scribus XSLT processing for handling frame-specific processing.
The key use-case now is to strip the first character of a text frame.
This can be used to handle the case where an image is used as a dropcap.

By adding an 'xslt-processing-configuration' attribute to a field, this option can be used.
Elements should be separated by semicolons, similar to conditions.
The field should get the value 'strip-first-character' as the first processing step.

In order to make this work, the attribute value is read and passed down to the subsequent XSLT templates.
On the deeper level the processing-configuration is check to see which processing should be applied.

Related issue: https://git.fsfe.org/FSFE/ada-zangemann/issues/421

Signed-off-by: Nico Rikken <nico.rikken@fsfe.org>